still ongoing: adding support of unevenly-spaced blades
