+++ packages/mcni/python/mcni/components/RadialCollimator.py
@@ -1,20 +1,11 @@
 #!/usr/bin/env python
 #
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+# Jiao Lin <jiao.lin@gmail.com>
 #
-#                                   Jiao Lin
-#                      California Institute of Technology
-#                      (C) 2007-2014  All Rights Reserved
-#
-# {LicenseText}
-#
-# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-#
-
 
 category = 'optics'
 
-
+import numpy as np
 from mcni.AbstractComponent import AbstractComponent
 
 class RadialCollimator( AbstractComponent ):
@@ -28,9 +19,8 @@ class RadialCollimator( AbstractComponent ):
         x = arr[:,0]; y = arr[:,1]; z = arr[:,2]
         vx = arr[:,3]; vy = arr[:,4]; vz = arr[:,5]
         s1 = arr[:,6]; s2 = arr[:,7];
-        t = arr[:,8]; 
+        t = arr[:,8];
         p = arr[:,9]
-        
         #
         theta_osc = self.oscillation * np.random.rand(len(neutrons))
         x1,y1,z1 = self.intersectCylinder(
@@ -42,33 +32,41 @@ class RadialCollimator( AbstractComponent ):
         good = (x1==x1) \
             * (z1<self.height1/2.) * (z1>-self.height1/2.) \
             * (z2<self.height2/2.) * (z2>-self.height2/2.) \
-            * (theta1 > self.theta1) * (theta1 < self.theta2) \
-            * (theta2 > self.theta1) * (theta2 < self.theta2) \
-            * ((theta1-self.theta1)//self.dtheta == (theta2-self.theta1)//self.dtheta)
+            * (theta1 > self.theta_min) * (theta1 < self.theta_max) \
+            * (theta2 > self.theta_min) * (theta2 < self.theta_max)
+        if self.dtheta is not None:
+            good *= (theta1-self.theta_min)//self.dtheta == (theta2-self.theta_min)//self.dtheta
+        else:
+            good *= np.digitize(theta1, self.theta_list) == np.digitize(theta1, self.theta_list)
         good = arr[good, :]
-        
         neutrons.resize(good.shape[0], neutrons[0])
         neutrons.from_npyarr(good)
         return
-    
-    
+
     def __init__(
-        self, name, 
-        radius1, height1, radius2, height2,
-        theta1, theta2, 
-        dtheta, oscillation=0):
+            self, name,
+            radius1, height1, radius2, height2,
+            theta1=0, theta2=np.pi, dtheta=np.radians(10),
+            theta_list = None,
+            oscillation=0
+    ):
         self.name = name
         self.radius1 = radius1
         self.radius2 = radius2
         self.height1 = height1
         self.height2 = height2
-        self.theta1 = theta1
-        self.theta2 = theta2
-        self.dtheta = dtheta
+        if theta1 is not None and theta2 is not None and dtheta is not None:
+            self.theta_min = theta1
+            self.theta_max = theta2
+            self.dtheta = dtheta
+        else:
+            self.theta_min = theta_list[0]
+            self.theta_max = theta_list[-1]
+            self.theta_list = theta_list
+            self.dtheta = None
         self.oscillation = oscillation
         return
-    
-    
+
     def intersectCylinder(self, position, velocity, cylinder):
         radius, height = cylinder
         x,y,z = position
@@ -78,61 +76,8 @@ class RadialCollimator( AbstractComponent ):
         C = x*x + y*y - radius*radius
         t = (np.sqrt(B*B-4*A*C) - B)/2/A
         return x+vx*t, y+vy*t, z+vz*t
-    
-    
+
     pass # end of MonochromaticSource
 
 
-import numpy as np
-
-
-def test():
-    DEG2RAD = np.pi/180
-    from mcni import neutron_buffer, neutron
-    neutrons = neutron_buffer(1)
-    coll = RadialCollimator(
-        name="collimator",
-        radius1=0.308, height1=0.6, radius2=0.462, height2=0.6,
-        theta1=-30*DEG2RAD, theta2=150*DEG2RAD, 
-        dtheta=1.6*DEG2RAD)
-    def check(neutron, absorbed):
-        neutrons.resize(1, neutron)
-        neutrons[0] = neutron
-        coll.process(neutrons)
-        assert len(neutrons) == (not absorbed)
-        return
-    
-    check(neutron(r=(0,0,0), v=(-0.9,0,1.732), prob=1), False)
-    check(neutron(r=(0,0,0), v=(-1.1,0,1.732), prob=1), True)
-    check(neutron(r=(0,0,0), v=(0,-0.3,0.463), prob=1), False)
-    check(neutron(r=(0,0,0), v=(0,-0.3,0.461), prob=1), True)
-    check(neutron(r=(0,0,0), v=(0,0.3,0.463), prob=1), False)
-    check(neutron(r=(0,0,0), v=(0,0.3,0.461), prob=1), True)
-    check(neutron(r=(0,0,0), v=(0.1,0.3,0.45), prob=1), True)
-    check(neutron(r=(0,0,0), v=(0.1,0.3,0.462), prob=1), False)
-    check(neutron(r=(0,0,0), v=(0,0,1000), prob=1), False)
-    check(neutron(r=(0,0,0), v=(1000,0,0), prob=1), False)
-    check(neutron(r=(0,0,0), v=(1000,0,-1000), prob=1), False)
-    check(neutron(r=(0,0,0), v=(0,0,-1000), prob=1), True)
-    check(neutron(r=(0.1,0,0), v=(0,0,1000), prob=1), True)
-    check(neutron(r=(0.001,0,0), v=(0,0,1000), prob=1), False)
-    check(neutron(r=(0.01,0,0), v=(0,0,1000), prob=1), False)
-    check(neutron(r=(0.03,0,0), v=(0,0,1000), prob=1), True)
-    check(neutron(r=(0.025,0,0), v=(0,0,1000), prob=1), True)
-    check(neutron(r=(0.015,0,0), v=(0,0,1000), prob=1), True)
-    check(neutron(r=(0.1,0,0), v=(0,10,0), prob=1), True)
-    for i in range(10):
-        check(neutron(r=(0,0,0), v=(1,0,i), prob=1), False)
-    for i in range(1,10):
-        check(neutron(r=(0,0,0), v=(i,0,-1), prob=1), False)
-        continue
-    return
-
-
-if __name__ == '__main__': test()
-    
-
-# version
-__id__ = "$Id$"
-
-# End of file 
+# End of file
